+++ client/src/Components/Article/ArticleDetail/index.tsx
@@ -1,8 +1,7 @@
 import { useQuery } from 'react-query';
 import { Link, useNavigate, useParams } from 'react-router-dom';
 import ArticleAPI from '../../../api/article';
-import tw from 'tailwind-styled-components';
-import styled from 'styled-components';
+import S from './styled';
 import ArticleWriterAndUpdatedAt from '../ArticleWriterAndUpdateAt';
 import { useStore } from '../../../store/store';
 import CommentContainer from '../../Comment/CommentContainer';
@@ -24,93 +23,35 @@ const ArticleDetail = () => {
   };
 
   return (
-    <ArticleContainer>
-      <ArticleHeader>
-        <ArticleTitle>{articleDetailData?.title}</ArticleTitle>
-        <ArticleHeaderColumn>
+    <S.ArticleContainer>
+      <S.ArticleHeader>
+        <S.ArticleTitle>{articleDetailData?.title}</S.ArticleTitle>
+        <S.ArticleHeaderColumn>
           <ArticleWriterAndUpdatedAt
             user={articleDetailData?.user!}
             updatedAt={String(articleDetailData?.updatedAt)}
           />
           {articleDetailData?.user.id === user?.id ? (
-            <ArticleDeleteBtn onClick={onDelete}>삭제</ArticleDeleteBtn>
+            <S.ArticleDeleteBtn onClick={onDelete}>삭제</S.ArticleDeleteBtn>
           ) : null}
-        </ArticleHeaderColumn>
-      </ArticleHeader>
-      <ArticleBodyContainer>
-        <ArticleTags>
+        </S.ArticleHeaderColumn>
+      </S.ArticleHeader>
+      <S.ArticleBodyContainer>
+        <S.ArticleTags>
           {tagList?.map((tag) => (
             <Link to={`/@${user?.username}?tag=${tag.substring(1)}`}>
-              <ArticleTag key={tag}>{tag}</ArticleTag>
+              <S.ArticleTag key={tag}>{tag}</S.ArticleTag>
             </Link>
           ))}
-        </ArticleTags>
-        <ArticleBody>{articleDetailData?.body}</ArticleBody>
+        </S.ArticleTags>
+        <S.ArticleBody>{articleDetailData?.body}</S.ArticleBody>
         <CommentContainer
           articleId={+articleId!}
           comments={articleDetailData?.comments!}
         />
-      </ArticleBodyContainer>
-    </ArticleContainer>
+      </S.ArticleBodyContainer>
+    </S.ArticleContainer>
   );
 };
-
-const ArticleContainer = tw.div`
-w-full
-flex
-flex-col
-`;
-const ArticleHeaderColor = styled.div`
-  color: ${(props) => props.theme.textColor};
-  background-color: ${(props) => props.theme.bannerColor};
-`;
-const ArticleHeader = tw(ArticleHeaderColor)`
-flex
-flex-col
-border
-border-solid
-px-[15%]
-pb-5
-mt-16
-`;
-
-const ArticleHeaderColumn = tw.div`
-flex
-justify-between
-w-full
-`;
-
-const ArticleTitle = tw.h1`
-mt-20
-text-4xl
-font-bold
-mx-auto
-`;
-
-const ArticleTags = tw.ul`
-mt-2
-flex
-gap-2
-`;
-const ArticleTag = tw.li`
-border
-border-solid
-px-2
-py-1
-rounded-md
-`;
-const ArticleBodyContainer = tw.div`
-w-[80%]
-mx-auto
-mt-2
-`;
-const ArticleBody = tw.p`
-mt-3
-`;
-
-const ArticleDeleteBtn = tw.button`
-
-
-`;
 
 export default ArticleDetail;
