+++ client/src/Components/Article/ArticleList/index.tsx
@@ -1,47 +1,17 @@
+import S from './styled';
 import { useQuery } from 'react-query';
-import styled from 'styled-components';
 import ArticleAPI from '../../../api/article';
-import tw from 'tailwind-styled-components';
 import { Link } from 'react-router-dom';
 import ArticleWriterAndUpdatedAt from '../ArticleWriterAndUpdateAt';
-
-const ArticleContainer = tw.div`
-w-full 
-`;
-
-const Articles = tw.ul`
-w-1/2
-mx-auto
-divide-y-2
-divide-solid
-`;
-const Article = styled.li`
-  padding: 20px 0;
-`;
-
-const ArticleTitleColor = styled.h2`
-  color: ${(props) => props.theme.textColor};
-`;
-const ArticleTitle = tw(ArticleTitleColor)`
-text-xl
-font-medium
-`;
-const ArticleDescriptonColor = styled.span`
-  color: ${(props) => props.theme.lightTextColor};
-`;
-const ArticleDescripton = tw(ArticleDescriptonColor)`
-text-lg
-font-light
-`;
 
 const ArticleList = () => {
   const { data: articleListData } = useQuery('articleList', ArticleAPI.getAll);
   return (
     <>
-      <ArticleContainer>
-        <Articles>
+      <S.ArticleContainer>
+        <S.Articles>
           {articleListData?.map((article) => (
-            <Article key={article.id}>
+            <S.Article key={article.id}>
               <ArticleWriterAndUpdatedAt
                 user={article.user}
                 updatedAt={article.updatedAt + ''}
@@ -50,13 +20,13 @@ const ArticleList = () => {
                 key={article.id}
                 to={`/@${article.user.username}/${article.id}`}
               >
-                <ArticleTitle>{article.title}</ArticleTitle>
+                <S.ArticleTitle>{article.title}</S.ArticleTitle>
               </Link>
-              <ArticleDescripton>{article.description}</ArticleDescripton>
-            </Article>
+              <S.ArticleDescripton>{article.description}</S.ArticleDescripton>
+            </S.Article>
           ))}
-        </Articles>
-      </ArticleContainer>
+        </S.Articles>
+      </S.ArticleContainer>
     </>
   );
 };
