+++ client/src/Components/Article/ArticleList2/index.tsx
@@ -1,39 +1,8 @@
 import { useQuery } from 'react-query';
 import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-import tw from 'tailwind-styled-components';
+import S from './styled';
 import ArticleAPI from '../../../api/article';
 import ArticleWriterAndUpdatedAt from '../ArticleWriterAndUpdateAt';
-
-const ArticleContainer = tw.div`
-w-full 
-mt-[100px]
-`;
-
-const Articles = tw.ul`
-w-1/2
-mx-auto
-divide-y-2
-divide-solid
-`;
-const Article = styled.li`
-  padding: 20px 0;
-`;
-
-const ArticleTitleColor = styled.h2`
-  color: ${(props) => props.theme.textColor};
-`;
-const ArticleTitle = tw(ArticleTitleColor)`
-text-xl
-font-medium
-`;
-const ArticleDescriptonColor = styled.span`
-  color: ${(props) => props.theme.lightTextColor};
-`;
-const ArticleDescripton = tw(ArticleDescriptonColor)`
-text-lg
-font-light
-`;
 
 interface ArticleList2Props {
   username: string;
@@ -47,10 +16,10 @@ const ArticleList2 = ({ username, tag }: ArticleList2Props) => {
   );
   return (
     <div>
-      <ArticleContainer>
-        <Articles>
+      <S.ArticleContainer>
+        <S.Articles>
           {articleListData?.map((article) => (
-            <Article key={article.id}>
+            <S.Article key={article.id}>
               <ArticleWriterAndUpdatedAt
                 user={article.user}
                 updatedAt={article.updatedAt + ''}
@@ -59,13 +28,13 @@ const ArticleList2 = ({ username, tag }: ArticleList2Props) => {
                 key={article.id}
                 to={`/@${article.user.username}/${article.id}`}
               >
-                <ArticleTitle>{article.title}</ArticleTitle>
+                <S.ArticleTitle>{article.title}</S.ArticleTitle>
               </Link>
-              <ArticleDescripton>{article.description}</ArticleDescripton>
-            </Article>
+              <S.ArticleDescripton>{article.description}</S.ArticleDescripton>
+            </S.Article>
           ))}
-        </Articles>
-      </ArticleContainer>
+        </S.Articles>
+      </S.ArticleContainer>
     </div>
   );
 };
