+++ client/src/Components/Article/ArticleWriterAndUpdateAt/index.tsx
@@ -1,26 +1,8 @@
 import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-import tw from 'tailwind-styled-components';
+import S from './styled';
 import useDate from '../../../hooks/common/useDate';
 import { IUser } from '../../../interfaces/user.interface';
 
-const ArticleWriterAndUpdatedAtContainer = tw.div`
-flex
-flex-col
-`;
-const ArticleWriterColor = styled.span`
-  color: ${(props) => props.theme.textColor};
-`;
-const ArticleWriter = tw(ArticleWriterColor)`
-text-base
-font-semibold
-`;
-const ArticleUpdatedAtColor = styled.span`
-  color: ${(props) => props.theme.lightTextColor};
-`;
-const ArticleUpdatedAt = tw(ArticleUpdatedAtColor)`
-text-xs
-`;
 interface IWriterAndUpdatedAt {
   user: IUser;
   updatedAt: string;
@@ -33,12 +15,12 @@ const ArticleWriterAndUpdatedAt = ({
   const articleUpdatedAt = useDate(updatedAt);
   return (
     <>
-      <ArticleWriterAndUpdatedAtContainer>
-        <ArticleWriter>
+      <S.ArticleWriterAndUpdatedAtContainer>
+        <S.ArticleWriter>
           <Link to={`/@${user.username}`}>{user.username}</Link>
-        </ArticleWriter>
-        <ArticleUpdatedAt>{articleUpdatedAt}</ArticleUpdatedAt>
-      </ArticleWriterAndUpdatedAtContainer>
+        </S.ArticleWriter>
+        <S.ArticleUpdatedAt>{articleUpdatedAt}</S.ArticleUpdatedAt>
+      </S.ArticleWriterAndUpdatedAtContainer>
     </>
   );
 };
