+++ client/src/Components/Auth/AuthModal/index.tsx
@@ -1,4 +1,4 @@
-import styled, { css } from 'styled-components';
+import S from './styled';
 
 interface AuthModalProps {
   visible: boolean;
@@ -8,31 +8,11 @@ interface AuthModalProps {
 
 const AuthModal = ({ visible, onClose }: AuthModalProps) => {
   return (
-    <AuthModalWrapper visible={visible}>
+    <S.AuthModalWrapper visible={visible}>
       <div>AuthModal</div>
       <button onClick={onClose}>버튼</button>
-    </AuthModalWrapper>
+    </S.AuthModalWrapper>
   );
 };
-
-const AuthModalWrapper = styled.div<{ visible: boolean }>`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 100;
-  ${(props) =>
-    props.visible
-      ? css`
-          background-color: rgba(0, 0, 0, 0.1);
-        `
-      : css`
-          visibility: hidden;
-        `}
-`;
 
 export default AuthModal;
