+++ client/src/Components/Auth/AuthModal/styled.ts
@@ -1,0 +1,25 @@
+import styled, { css } from 'styled-components';
+
+const AuthModalWrapper = styled.div<{ visible: boolean }>`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 100;
+  ${(props) =>
+    props.visible
+      ? css`
+          background-color: rgba(0, 0, 0, 0.1);
+        `
+      : css`
+          visibility: hidden;
+        `}
+`;
+
+export default {
+  AuthModalWrapper,
+};
