+++ client/src/Components/Comment/CommentContainer/styled.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const CommentContainerBlock = styled.div`
+  margin-top: 2rem;
+`;
+
+export default {
+  CommentContainerBlock,
+};
