+++ client/src/Components/Comment/CommentItem/index.tsx
@@ -1,10 +1,10 @@
-import styled from 'styled-components';
+import S from './styled';
 import { IComment } from '../../../interfaces/comment.interface';
 import ArticleWriterAndUpdatedAt from '../../Article/ArticleWriterAndUpdateAt';
-import tw from 'tailwind-styled-components';
 import CommentAPI from '../../../api/comment';
 import { useStore } from '../../../store/store';
 import { useMutation, useQueryClient } from 'react-query';
+
 interface CommentItemProps {
   comment: IComment;
   articleId: number;
@@ -28,50 +28,19 @@ const CommentItem = ({ comment, articleId }: CommentItemProps) => {
   );
 
   return (
-    <Card>
-      <CardBlock>{comment.text}</CardBlock>
-      <CardFooter>
+    <S.Card>
+      <S.CardBlock>{comment.text}</S.CardBlock>
+      <S.CardFooter>
         <ArticleWriterAndUpdatedAt
           user={comment.user}
           updatedAt={comment.updatedAt + ''}
         />
         {comment.user.id === user?.id && (
-          <CommentDeleteBtn onClick={onCommentDelete}>❌</CommentDeleteBtn>
+          <S.CommentDeleteBtn onClick={onCommentDelete}>❌</S.CommentDeleteBtn>
         )}
-      </CardFooter>
-    </Card>
+      </S.CardFooter>
+    </S.Card>
   );
 };
-
-const Card = styled('div')`
-  margin-bottom: 0.75rem;
-  background-color: #fff;
-  border-radius: 0.25rem;
-  border: 1px solid #e5e5e5;
-`;
-
-const CardBlock = styled('div')`
-  width: 100%;
-  font-size: 1rem;
-  line-height: 1.25;
-  background-color: #fff;
-  border-radius: 0.25rem;
-  border: 0;
-  padding: 1.25rem;
-  outline: none;
-  color: black;
-`;
-
-const CardFooter = styled('div')`
-  border-top: 1px solid #e5e5e5;
-  font-size: 0.8rem;
-  font-weight: 300;
-  padding: 0.75rem 1.25rem;
-  background-color: #f5f5f5;
-`;
-
-const CommentDeleteBtn = tw.span`
-hover:cursor-pointer
-`;
 
 export default CommentItem;
